Update README.md
In this project we are going to build a project based on Library Management System. It keeps track of all information about books in the library, their cost, status and total number of books available in the library.
Here the tables contains :
1. Branch 
2. Employee 
3. Books
4. Customer
5. IssueStatus
6. ReturnStatus 
The Library Management System project successfully created a robust database infrastructure that supports efficient library operations and provides a solid foundation for future enhancements. This project demonstrates the critical role of well-designed databases in managing and optimizing complex information systems.